Remove `User` (full user) + add extra people to `UserDTO`
